Add duplicate action to filter drawer
