+++ packages/builder/src/components/design/PropertiesPanel/PropertyControls/FilterEditor/FilterEditor.svelte
@@ -1,19 +1,11 @@
 <script>
-  import {
-    notifications,
-    ActionButton,
-    Button,
-    Drawer,
-    Body,
-    DrawerContent,
-    Layout,
-  } from "@budibase/bbui"
+  import { notifications, ActionButton, Button, Drawer } from "@budibase/bbui"
   import { createEventDispatcher } from "svelte"
   import {
     getDatasourceForProvider,
     getSchemaForDatasource,
   } from "builderStore/dataBinding"
-  import LuceneFilterBuilder from "./LuceneFilterBuilder.svelte"
+  import FilterDrawer from "./FilterDrawer.svelte"
   import { currentAsset } from "builderStore"
 
   const dispatch = createEventDispatcher()
@@ -25,9 +17,6 @@
   let drawer
   let tempValue = value || []
 
-  $: numFilters = Array.isArray(tempValue)
-    ? tempValue.length
-    : Object.keys(tempValue || {}).length
   $: dataSource = getDatasourceForProvider($currentAsset, componentInstance)
   $: schema = getSchemaForDatasource($currentAsset, dataSource)?.schema
   $: schemaFields = Object.values(schema || {})
@@ -43,17 +32,10 @@
 <ActionButton on:click={drawer.show}>Define filters</ActionButton>
 <Drawer bind:this={drawer} title="Filtering">
   <Button cta slot="buttons" on:click={saveFilter}>Save</Button>
-  <DrawerContent slot="body">
-    <Layout noPadding>
-      <Body size="S">
-        {#if !numFilters}
-          Add your first filter column.
-        {:else}
-          Results are filtered to only those which match all of the following
-          constraints.
-        {/if}
-      </Body>
-      <LuceneFilterBuilder bind:value={tempValue} {schemaFields} {bindings} />
-    </Layout>
-  </DrawerContent>
+  <FilterDrawer
+    slot="body"
+    bind:filters={tempValue}
+    {bindings}
+    {schemaFields}
+  />
 </Drawer>
